requirements.txt
SQLAlchemy==2.0.45
pyodbc==5.3.0
tabulate==0.9.0
python-dotenv==1.0.0
pandas==2.3.3
mypy==1.19.1
Flask==3.0.0
openpyxl==3.1.2